fix: Moved all invidiual functions files from trading into index and made new https file
