+++ app/src/util/tools/fpmm/index.ts
@@ -1,13 +1,3 @@
-export {
-  calcPoolTokens,
-  calcAddFundingSendAmounts,
-  calcDepositedTokens,
-  calcRemoveFundingSendAmounts,
-  calcDistributionHint,
-  calcInitialFundingSendAmounts,
-  calcAddFundingDepositedAmounts,
-  calcInitialFundingDepositedAmounts,
-  calcSharesBought,
-} from './liquidity'
-export { calcPrice } from './price'
+export * from './liquidity'
+export * from './price'
 export * from './trading'
